fixed bug

+++ ScriptedCalculations/setup_filestructure.py
@@ -82,8 +82,10 @@ def setup_esp_filestructure(structfilename, env_config, basisfilename, workdir,
     espdir = compounddir/'esp'
 
     # Check for calculation dir in working directory
+    existing_directory = False
     if compounddir.exists():
         print("Found existing working directory.")
+        existing_directory = True
     else:
         # If directory does not exist, make it
         os.mkdir(compounddir)
@@ -91,7 +93,7 @@ def setup_esp_filestructure(structfilename, env_config, basisfilename, workdir,
 
     # If optimized and centered xyz file in input_xyz directory
     optimized_xyz = 'input_xyz/{}_optimized_centered.xyz'.format(compoundname)
-    if optimized_xyz.exists():
+    if optimized_xyz.exists() and not existing_directory:
         print("Found optimized gemoetry file.")
         shutil.copy(optimized_xyz, compounddir/'gndstate')
     else:
